# docs/swagger.yaml
definitions:
  internal_handler.UserCreateReq:
    properties:
      age:
        maximum: 120
        minimum: 1
        type: integer
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - nickname
    - password
    - username
    type: object
  internal_handler.UserListItem:
    properties:
      email:
        description: 邮箱
        type: string
      id:
        description: 用户ID
        type: integer
      nickname:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        example: username
        type: string
    type: object
  internal_handler.UserListRes:
    properties:
      list:
        items:
          $ref: '#/definitions/internal_handler.UserListItem'
        type: array
      total:
        type: integer
    type: object
  scaffold_common_httpx.Resp:
    properties:
      code:
        description: 业务错误码,正常返回为200，错误返回为错误码
        type: integer
      data:
        description: 返回数据
      msg:
        description: 错误信息，如果状态码为200，msg为OK
        type: string
    type: object
info:
  contact: {}
paths:
  /user:
    get:
      consumes:
      - application/json
      description: 用户列表
      parameters:
      - description: 年龄
        example: 101
        in: query
        name: age
        type: integer
      - description: 生日
        example: "2020-01-01"
        in: query
        name: birthday
        type: string
      - description: 邮箱
        example: zhangsanfeng@163.com
        in: query
        name: email
        type: string
      - description: 手机号
        example: "13888888888"
        in: query
        name: mobile
        type: string
      - description: 昵称
        example: 张三疯
        in: query
        name: nickname
        type: string
      - description: 用户名
        example: 张三
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 响应结果
          schema:
            allOf:
            - $ref: '#/definitions/scaffold_common_httpx.Resp'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.UserListRes'
              type: object
      summary: 用户列表
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 新增用户
      parameters:
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应结果
          schema:
            $ref: '#/definitions/scaffold_common_httpx.Resp'
      summary: 新增用户
      tags:
      - user
swagger: "2.0"